requirements.txt
astroid==2.3.3
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
pylint==2.4.4
pynput==1.6.8
python-docx==0.8.10
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
